Enrich API with Doc and Token
Current wheel fail with undefined symbol.
