+++ Dockerfile-manylinux_2_28
@@ -40,7 +40,6 @@ ENV PATH=/opt/_internal/cpython-${PYTHON_FULL_VERSION}/bin:/opt/python/${PYTHON_
     PYTHON_FULL_VERSION=$PYTHON_FULL_VERSION \
     PYTHON_WHEEL_VERSION=$PYTHON_WHEEL_VERSION \
     LIMA_PYTHON_VERSION=$LIMA_PYTHON_VERSION \
-    CLANG_INSTALL_DIR=/usr/lib \
     Qt_INSTALL_DIR=/opt/qt6
 
 # Build LIMA wheel
@@ -53,10 +52,10 @@ RUN cp -R /opt/lima/share/config/lima/* /lima-python/aymaralima/config
 
 RUN python${PYTHON_VERSION} /lima-python/scripts/linuxdeploy.py /opt/lima/lib/liblima*.so -d clib -o clib/libs.json
 RUN /usr/bin/strip --strip-unneeded clib/lib/*.so
-COPY scripts tests utils bindings.h bindings.xml c2lc.txt CHANGES.md CMakeLists.txt deploy.sh LICENCE lima.h lima.cpp Doc.h Doc.cpp Token.h Token.cpp macros.h manageQt.cmake MANIFEST.in python_env.sh README.md setup.py /lima-python/
+COPY scripts tests utils bindings.h bindings.xml c2lc.txt CHANGES.md CMakeLists.txt deploy.sh LICENCE lima.h lima.cpp Doc.h Doc_private.h Doc.cpp Token.h Token.cpp macros.h manageQt.cmake MANIFEST.in python_env.sh README.md setup.py /lima-python/
 
 COPY aymara/* /lima-python/aymara
-
+ARG CACHE_BUST=1
 RUN python${PYTHON_VERSION} setup.py bdist_wheel
 ENV LD_LIBRARY_PATH=/lima-python/_skbuild/linux-x86_64-${PYTHON_VERSION}/cmake-build:$LD_LIBRARY_PATH
 # RUN auditwheel show /lima-python/dist/aymara-${LIMA_PYTHON_VERSION}-${PYTHON_SHORT_VERSION}-linux_x86_64.whl
