+++ Token.h
@@ -58,7 +58,6 @@
 
 #include "macros.h"
 
-#include <memory>
 #include <string>
 #include <vector>
